feat(android): fix screen sharing for android 14 (#14419)
* feat(android): media projection is now done through react native webrtc
