+++ android/build.gradle
@@ -19,9 +19,9 @@ buildscript {
 ext {
     kotlinVersion = "1.7.0"
     buildToolsVersion = "33.0.2"
-    compileSdkVersion = 33
+    compileSdkVersion = 34
     minSdkVersion    = 24
-    targetSdkVersion = 33
+    targetSdkVersion = 34
     supportLibVersion = "28.0.0"
 
     // We use NDK 23 which has both M1 support and is the side-by-side NDK version from AGP.
