--- android/sdk/src/main/java/org/jitsi/meet/sdk/JitsiMeetMediaProjectionModule.java
@@ -1,44 +1,0 @@
-package org.jitsi.meet.sdk;
-
-import android.app.Activity;
-import android.content.Context;
-
-import androidx.annotation.NonNull;
-
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.bridge.ReactContextBaseJavaModule;
-import com.facebook.react.bridge.ReactMethod;
-import com.facebook.react.module.annotations.ReactModule;
-
-
-@ReactModule(name = JitsiMeetMediaProjectionModule.NAME)
-class JitsiMeetMediaProjectionModule
-    extends ReactContextBaseJavaModule {
-
-    public static final String NAME = "JitsiMeetMediaProjectionModule";
-
-    public JitsiMeetMediaProjectionModule(ReactApplicationContext reactContext) {
-        super(reactContext);
-    }
-
-    @ReactMethod
-    public void launch() {
-        Context context = getReactApplicationContext();
-        Activity currentActivity = getCurrentActivity();
-
-        JitsiMeetMediaProjectionService.launch(context, currentActivity);
-    }
-
-    @ReactMethod
-    public void abort() {
-        Context context = getReactApplicationContext();
-
-        JitsiMeetMediaProjectionService.abort(context);
-    }
-
-    @NonNull
-    @Override
-    public String getName() {
-        return NAME;
-    }
-}
