--- android/sdk/src/main/java/org/jitsi/meet/sdk/JitsiMeetMediaProjectionService.java
@@ -1,102 +1,0 @@
-/*
- * Copyright @ 2019-present 8x8, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jitsi.meet.sdk;
-
-
-import android.app.Activity;
-import android.app.Notification;
-import android.app.Service;
-import android.content.ComponentName;
-import android.content.Context;
-import android.content.Intent;
-import android.content.pm.ServiceInfo;
-import android.os.Build;
-import android.os.IBinder;
-
-import org.jitsi.meet.sdk.log.JitsiMeetLogger;
-
-import java.util.Random;
-
-/**
- * This class implements an Android {@link Service}, a foreground one specifically, and it's
- * responsible for presenting an ongoing notification when a conference is in progress.
- * The service will help keep the app running while in the background.
- *
- * See: https://developer.android.com/guide/components/services
- */
-public class JitsiMeetMediaProjectionService extends Service {
-    private static final String TAG = JitsiMeetMediaProjectionService.class.getSimpleName();
-
-    static final int NOTIFICATION_ID = new Random().nextInt(99999) + 10000;
-
-    public static void launch(Context context, Activity currentActivity) {
-
-        NotificationUtils.createNotificationChannel(currentActivity);
-
-        Intent intent = new Intent(context, JitsiMeetMediaProjectionService.class);
-
-        ComponentName componentName;
-
-        try {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                componentName = context.startForegroundService(intent);
-            } else {
-                componentName = context.startService(intent);
-            }
-        } catch (RuntimeException e) {
-            // Avoid crashing due to ForegroundServiceStartNotAllowedException (API level 31).
-            // See: https://developer.android.com/guide/components/foreground-services#background-start-restrictions
-            JitsiMeetLogger.w(TAG + "Media projection service not started", e);
-            return;
-        }
-
-        if (componentName == null) {
-            JitsiMeetLogger.w(TAG + "Media projection service not started");
-        }
-    }
-
-    public static void abort(Context context) {
-        Intent intent = new Intent(context, JitsiMeetMediaProjectionService.class);
-        context.stopService(intent);
-    }
-
-    @Override
-    public IBinder onBind(Intent intent) {
-        return null;
-    }
-
-    @Override
-    public int onStartCommand(Intent intent, int flags, int startId) {
-
-        Notification notification = MediaProjectionNotification.buildMediaProjectionNotification(this);
-
-        if (notification == null) {
-            stopSelf();
-            JitsiMeetLogger.w(TAG + " Couldn't start service, notification is null");
-
-            return START_NOT_STICKY;
-        }
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            startForeground(NOTIFICATION_ID, notification, ServiceInfo.FOREGROUND_SERVICE_TYPE_MEDIA_PROJECTION);
-        } else {
-            startForeground(NOTIFICATION_ID, notification);
-        }
-
-        return START_NOT_STICKY;
-    }
-}
