+++ android/sdk/src/main/java/org/jitsi/meet/sdk/OngoingNotification.java
@@ -16,8 +16,11 @@
 
 package org.jitsi.meet.sdk;
 
-import static org.jitsi.meet.sdk.NotificationUtils.ONGOING_CONFERENCE_CHANNEL_ID;
+import org.jitsi.meet.sdk.log.JitsiMeetLogger;
 
+import android.app.Activity;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
 import android.app.Notification;
 import android.app.PendingIntent;
 import android.content.Context;
@@ -26,7 +29,8 @@ import android.content.Intent;
 import androidx.annotation.StringRes;
 import androidx.core.app.NotificationCompat;
 
-import org.jitsi.meet.sdk.log.JitsiMeetLogger;
+import android.os.Build;
+
 
 /**
  * Helper class for creating the ongoing notification which is used with
@@ -37,6 +41,37 @@ class OngoingNotification {
     private static final String TAG = OngoingNotification.class.getSimpleName();
 
     private static long startingTime = 0;
+
+    static final String ONGOING_CONFERENCE_CHANNEL_ID = "JitsiOngoingConferenceChannel";
+
+    static void createNotificationChannel(Activity context) {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O) {
+            return;
+        }
+
+        if (context == null) {
+            JitsiMeetLogger.w(TAG + " Cannot create notification channel: no current context");
+            return;
+        }
+
+        NotificationManager notificationManager
+            = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
+
+        NotificationChannel channel
+            = notificationManager.getNotificationChannel(ONGOING_CONFERENCE_CHANNEL_ID);
+
+        if (channel != null) {
+            // The channel was already created, no need to do it again.
+            return;
+        }
+
+        channel = new NotificationChannel(ONGOING_CONFERENCE_CHANNEL_ID, context.getString(R.string.ongoing_notification_channel_name), NotificationManager.IMPORTANCE_DEFAULT);
+        channel.enableLights(false);
+        channel.enableVibration(false);
+        channel.setShowBadge(false);
+
+        notificationManager.createNotificationChannel(channel);
+    }
 
     static Notification buildOngoingConferenceNotification(Boolean isMuted, Context context) {
